requirements.txt
streamlit==1.30.0
pandas==2.1.3
python-dotenv==1.0.0